Version 3 of the codework
This is the version of my work in the competition. A bunch of EDA steps have been added to this work. 
Regex pattern matching was the main addition in this update. Apart from the brand & model columns, all other columns were handled. A few numerical columns were newly added. 
More updates to come.